requirements.txt
Django==1.7
djangorestframework
factory_boy
